LayoutTests/fast/forms/ios/accessory-bar-navigation-expected.txt

Tests zooming into a text input on tap.


After focus first field

scale	0.750
visibleRect	{ left: 0, top: 0, width: 520, height: 1063 }
After move to second field

scale	0.750
visibleRect	{ left: 0, top: 766, width: 520, height: 1063 }
After move to first field

scale	0.750
visibleRect	{ left: 0, top: 0, width: 520, height: 1063 }
